Rename project to c2puml, remove unused modules, update README
