--- main.py
@@ -1,11 +1,0 @@
-#!/usr/bin/env python3
-"""
-C to PlantUML Converter - Main Entry Point
-"""
-
-import sys
-
-from c_to_plantuml.main import main
-
-if __name__ == "__main__":
-    sys.exit(main())
