--- output/diagram_index.html
@@ -1,219 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>PlantUML Diagrams - C to PlantUML Converter</title>
-    <style>
-        body { 
-            font-family: -apple-system, BlinkMacSystemFont, 'Segoe UI', Roboto, sans-serif; 
-            line-height: 1.6; 
-            margin: 0; 
-            padding: 20px; 
-            background: #f6f8fa; 
-        }
-        .container { 
-            max-width: 1200px; 
-            margin: 0 auto; 
-        }
-        .header { 
-            background: white; 
-            padding: 20px; 
-            border-radius: 6px; 
-            margin-bottom: 20px; 
-            box-shadow: 0 1px 3px rgba(0,0,0,0.1); 
-        }
-        .nav { 
-            background: #0366d6; 
-            color: white; 
-            padding: 15px 20px; 
-            border-radius: 6px; 
-            margin-bottom: 20px; 
-        }
-        .nav a { 
-            color: white; 
-            text-decoration: none; 
-            margin-right: 20px; 
-            font-weight: 500; 
-        }
-        .nav a:hover { 
-            text-decoration: underline; 
-        }
-        .stats { 
-            background: white; 
-            padding: 15px; 
-            border-radius: 6px; 
-            margin-bottom: 20px; 
-            box-shadow: 0 1px 3px rgba(0,0,0,0.1); 
-        }
-        .diagram-grid { 
-            display: grid; 
-            grid-template-columns: repeat(auto-fill, minmax(400px, 1fr)); 
-            gap: 20px; 
-            margin-top: 20px; 
-        }
-        .diagram-card { 
-            background: white; 
-            border-radius: 6px; 
-            box-shadow: 0 1px 3px rgba(0,0,0,0.1); 
-            overflow: hidden; 
-            transition: transform 0.2s; 
-        }
-        .diagram-card:hover { 
-            transform: translateY(-2px); 
-            box-shadow: 0 4px 8px rgba(0,0,0,0.15); 
-        }
-        .diagram-card h3 { 
-            margin: 0; 
-            padding: 15px; 
-            background: #f1f3f4; 
-            border-bottom: 1px solid #e1e4e8; 
-            font-size: 16px; 
-            color: #24292e; 
-        }
-        .diagram-image { 
-            width: 100%; 
-            height: auto; 
-            display: block; 
-        }
-        .diagram-links { 
-            padding: 15px; 
-            background: #f8f9fa; 
-        }
-        .diagram-links a { 
-            display: inline-block; 
-            margin: 5px 10px 5px 0; 
-            padding: 5px 10px; 
-            background: #0366d6; 
-            color: white; 
-            text-decoration: none; 
-            border-radius: 4px; 
-            font-size: 12px; 
-        }
-        .diagram-links a:hover { 
-            background: #0256b3; 
-        }
-        .footer { 
-            text-align: center; 
-            margin-top: 40px; 
-            padding: 20px; 
-            color: #586069; 
-        }
-        .no-diagrams { 
-            background: white; 
-            padding: 40px; 
-            text-align: center; 
-            border-radius: 6px; 
-            box-shadow: 0 1px 3px rgba(0,0,0,0.1); 
-        }
-    </style>
-</head>
-<body>
-    <div class="container">
-        <div class="nav">
-            <span>📊 Diagrams</span>
-        </div>
-        
-        <div class="header">
-            <h1>📊 PlantUML Diagrams</h1>
-            <p>Generated diagrams from C/C++ source code analysis</p>
-            <p><strong>Generated:</strong> 2025-07-31 06:32:06 UTC</p>
-        </div>
-        
-        <div class="stats">
-            <h3>📈 Statistics</h3>
-            <p><strong>PNG Images:</strong> 9</p>
-            <p><strong>PlantUML Files:</strong> 9</p>
-            <p><strong>Total Files:</strong> 18</p>
-        </div>
-        <div class="diagram-grid">
-            
-            <div class="diagram-card">
-                <h3>complex</h3>
-                <img src="complex.png" alt="complex diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="complex.png" target="_blank">🖼️ Full Size</a>
-                    <a href="complex.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>geometry</h3>
-                <img src="geometry.png" alt="geometry diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="geometry.png" target="_blank">🖼️ Full Size</a>
-                    <a href="geometry.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>logger</h3>
-                <img src="logger.png" alt="logger diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="logger.png" target="_blank">🖼️ Full Size</a>
-                    <a href="logger.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>math_utils</h3>
-                <img src="math_utils.png" alt="math_utils diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="math_utils.png" target="_blank">🖼️ Full Size</a>
-                    <a href="math_utils.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>preprocessed</h3>
-                <img src="preprocessed.png" alt="preprocessed diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="preprocessed.png" target="_blank">🖼️ Full Size</a>
-                    <a href="preprocessed.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>sample</h3>
-                <img src="sample.png" alt="sample diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="sample.png" target="_blank">🖼️ Full Size</a>
-                    <a href="sample.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>sample2</h3>
-                <img src="sample2.png" alt="sample2 diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="sample2.png" target="_blank">🖼️ Full Size</a>
-                    <a href="sample2.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>transformed</h3>
-                <img src="transformed.png" alt="transformed diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="transformed.png" target="_blank">🖼️ Full Size</a>
-                    <a href="transformed.puml">📄 Source</a>
-                </div>
-            </div>
-            
-            <div class="diagram-card">
-                <h3>typedef_test</h3>
-                <img src="typedef_test.png" alt="typedef_test diagram" class="diagram-image" loading="lazy">
-                <div class="diagram-links">
-                    <a href="typedef_test.png" target="_blank">🖼️ Full Size</a>
-                    <a href="typedef_test.puml">📄 Source</a>
-                </div>
-            </div>
-        </div>
-        
-        <div class="footer">
-            <p>Generated by <a href="https://github.com/fischerjooo/generator_project">C to PlantUML Converter</a></p>
-            <p>Last updated: 2025-07-31 06:32:06 UTC</p>
-        </div>
-    </div>
-</body>
-</html>
