+++ run_example_fixed.bat
@@ -13,17 +13,17 @@ py -3.12 main.py --config example/config.json --verbose
 if %errorlevel% equ 0 goto :success
 
 REM Option 2: Try python3
-python3 main.py --config example/config.json --verbose
+python3 -m c2puml.main --config example/config.json --verbose
 if %errorlevel% equ 0 goto :success
 
 REM Option 3: Try python with full path
-"C:\toolbase\python\3.12.5.0-2\python.exe" main.py --config example/config.json --verbose
+"C:\toolbase\python\3.12.5.0-2\python.exe" -m c2puml.main --config example/config.json --verbose
 if %errorlevel% equ 0 goto :success
 
 REM Option 4: Try virtual environment
 if exist venv\Scripts\activate.bat (
     call venv\Scripts\activate.bat
-    python main.py --config example/config.json --verbose
+    python -m c2puml.main --config example/config.json --verbose
     if %errorlevel% equ 0 goto :success
 )
