+++ run_example_with_coverage.py
@@ -55,7 +55,7 @@ def run_example_generation():
             "run",
             "-a",
             "-m",
-            "c_to_plantuml.main",
+            "c2puml.main",
             "--config",
             "example/config.json",
         ],
@@ -65,7 +65,7 @@ def run_example_generation():
             "run",
             "-a",
             "-m",
-            "c_to_plantuml.main",
+            "c2puml.main",
             "--config",
             "example/config.json",
             "parse",
@@ -75,7 +75,7 @@ def run_example_generation():
             "run",
             "-a",
             "-m",
-            "c_to_plantuml.main",
+            "c2puml.main",
             "--config",
             "example/config.json",
             "transform",
@@ -85,7 +85,7 @@ def run_example_generation():
             "run",
             "-a",
             "-m",
-            "c_to_plantuml.main",
+            "c2puml.main",
             "--config",
             "example/config.json",
             "generate",
@@ -96,7 +96,7 @@ def run_example_generation():
             "run",
             "-a",
             "-m",
-            "c_to_plantuml.main",
+            "c2puml.main",
             "--config",
             "example/config.json",
             "--verbose",
