+++ tests/feature/test_crypto_filter_pattern.py
@@ -17,8 +17,8 @@ import tempfile
 import unittest
 from pathlib import Path
 
-from c_to_plantuml.config import Config
-from c_to_plantuml.parser import CParser
+from c2puml.config import Config
+from c2puml.parser import CParser
 from tests.feature.base import BaseFeatureTest
