+++ tests/unit/test_generator.py
@@ -13,8 +13,8 @@ from pathlib import Path
 
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))
 
-from c_to_plantuml.generator import Generator
-from c_to_plantuml.models import (
+from c2puml.generator import Generator
+from c2puml.models import (
     Alias,
     Enum,
     Field,
@@ -24,7 +24,7 @@ from c_to_plantuml.models import (
     Struct,
     Union,
 )
-from c_to_plantuml.parser import CParser
+from c2puml.parser import CParser
 
 
 class TestGenerator(unittest.TestCase):
