+++ tests/unit/test_include_processing.py
@@ -17,8 +17,8 @@ from unittest import TestCase
 
 sys.path.insert(0, os.path.join(os.path.dirname(__file__), ".."))
 
-from c_to_plantuml.generator import Generator
-from c_to_plantuml.parser import CParser
+from c2puml.generator import Generator
+from c2puml.parser import CParser
 
 
 class TestIncludeProcessingBug(TestCase):
