+++ tests/unit/test_preprocessor_bug.py
@@ -9,8 +9,8 @@ import tempfile
 import unittest
 from pathlib import Path
 
-from c_to_plantuml.parser import CParser
-from c_to_plantuml.parser_tokenizer import CTokenizer, TokenType
+from c2puml.parser import CParser
+from c2puml.parser_tokenizer import CTokenizer, TokenType
 
 
 class TestPreprocessorBug(unittest.TestCase):
