+++ tests/unit/test_preprocessor_handling.py
@@ -11,8 +11,8 @@ import tempfile
 import unittest
 from pathlib import Path
 
-from c_to_plantuml.parser import CParser
-from c_to_plantuml.parser_tokenizer import CTokenizer, TokenType
+from c2puml.parser import CParser
+from c2puml.parser_tokenizer import CTokenizer, TokenType
 
 
 class TestPreprocessorHandling(unittest.TestCase):
